fix: address PR drilldown follow-ups
- guard shift-compare against detached detail panels
- preserve selected term highlighting during hash restore
- split plo trend drilldown tests to satisfy code-sprawl gate
