+++ .secrets.baseline
@@ -143,5 +143,5 @@
       }
     ]
   },
-  "generated_at": "2026-04-03T17:58:53Z"
+  "generated_at": "2026-04-03T19:00:55Z"
 }
